# Optimize data flushing for `FileAccessCompressed` and `FileAccessEncrypted`