Improves the performance of  `FileAccessCompressed` and `FileAccessEncrypted`'s `_close()` functions by avoiding repeated allocations and COW.

Performance:
A file with 1000000 x int64, with each value being the index of the loop iteration.

| Before | After: |
|--------|--------|
| File Access Compressed: ~127 ms, File Access Encrypted: ~238 ms | File Access Compressed: ~120 ms, File Access Encrypted: ~207 ms |